refactor: change folder name
