+++ backend/src/main/factories/usecases/load-patrimony-by-code/load-patrimony-by-code.ts
